Change src directory
